+++ tests/Config/ConfigFile/Reader/ConfigJsonReaderTest.php
@@ -40,9 +40,9 @@ class ConfigJsonReaderTest extends \PHPUnit_Framework_TestCase
         $config = $configFile->getConfig();
         $this->assertSame('puli-dir', $config->get(Config::PULI_DIR));
         $this->assertSame('puli-dir/my-install-file.json', $config->get(Config::INSTALL_FILE));
-        $this->assertSame('puli-dir/my-repository.php', $config->get(Config::REPO_FILE));
-        $this->assertSame('puli-dir/my-repo', $config->get(Config::REPO_DUMP_DIR));
-        $this->assertSame('puli-dir/my-repository-dump.php', $config->get(Config::REPO_DUMP_FILE));
+        $this->assertSame('puli-dir/my-repository.php', $config->get(Config::READ_REPO));
+        $this->assertSame('puli-dir/my-repo', $config->get(Config::DUMP_DIR));
+        $this->assertSame('puli-dir/my-repository-dump.php', $config->get(Config::WRITE_REPO));
     }
 
     public function testReadMinimalConfigFile()
@@ -55,9 +55,9 @@ class ConfigJsonReaderTest extends \PHPUnit_Framework_TestCase
         $config = $configFile->getConfig();
         $this->assertSame('.puli', $config->get(Config::PULI_DIR));
         $this->assertSame('.puli/install-file.json', $config->get(Config::INSTALL_FILE));
-        $this->assertSame('.puli/resource-repository.php', $config->get(Config::REPO_FILE));
-        $this->assertSame('.puli/repo', $config->get(Config::REPO_DUMP_DIR));
-        $this->assertSame('.puli/resource-repository.php', $config->get(Config::REPO_DUMP_FILE));
+        $this->assertSame('.puli/resource-repository.php', $config->get(Config::READ_REPO));
+        $this->assertSame('.puli/dump', $config->get(Config::DUMP_DIR));
+        $this->assertSame('.puli/resource-repository.php', $config->get(Config::WRITE_REPO));
     }
 
     /**